fix bugs

+++ src/pages/WorkOrder/Diagram/index.tsx
@@ -1,7 +1,8 @@
-import { Autocomplete, Box, Collapse, Dialog, IconButton, TextField, Typography } from '@mui/material';
+import { Autocomplete, Box, Collapse, Dialog, Divider, IconButton, TextField, Typography } from '@mui/material';
 import { Add, Delete } from '@mui/icons-material';
 import EditIcon from '@mui/icons-material/Edit';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
+import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined';
 import KeyboardArrowRight from '@mui/icons-material/KeyboardArrowRight';
 import { useCallback, useContext, useEffect, useState } from 'react';
 import { isMobile, isTablet } from 'react-device-detect';
@@ -15,12 +16,14 @@ import CustomDialogHeader from 'src/components/CustomDialog/CustomDialogHeader';
 import HtmlTooltip from 'src/components/CustomTooltipTitle';
 import { ThemeButton } from 'src/components/Helpers/Buttons';
 import ConfirmationDialog from 'src/components/Helpers/ConfirmationDialog';
-import { ACTIVITY_RESOURCE, cn, CustomDialogTransition, WORK_ORDER_TYPE, workOrder } from 'src/constants/helpers';
+import { ACTIVITY_RESOURCE, cn, CustomDialogTransition, displayDate, WORK_ORDER_TYPE, workOrder } from 'src/constants/helpers';
 import PdfPreview from './ShowPdf/PdfPreview';
 import { getFileIcon, getFileNameWithExtension } from './utils';
 import emptyIllustration from 'src/assets/emptyIllustration.webp';
 import ImageEditor from './ImageEditor';
 import { useData } from 'src/StateProvider/Provider';
+import Comment from 'src/pages/WorkOrder/Diagram/Comment';
+import { isEmpty } from 'lodash';
 
 const imageExtensions = ['tif', 'tiff', 'bmp', 'jpg', 'jpeg', 'gif', 'png', 'eps', 'raw', 'cr2', 'nef', 'orf', 'sr2'];
 
@@ -49,13 +52,14 @@ const Diagram = ({
 
   const [rowData, setRowData] = useState(null);
   const [expended, setExpended] = useState({});
-  const [attachemntDialog, setAttachemntDialog] = useState({ open: false, id: null, isClone: false });
+  const [attachemntDialog, setAttachemntDialog] = useState({ open: false, file: null, isClone: false });
   const [fullScreen, setFullScreen] = useState(isMobile || isTablet);
   const [selectedFile, setSelectedFile] = useState(null);
   const [selectedAttachment, setSelectedAttachment] = useState(null);
   const [showConfirmBox, setShowConfirmBox] = useState(false);
   const [serviceOption, setServiceOption] = useState([]);
   const [selectedService, setSelectedService] = useState(null);
+  const [commentDialog, setCommentDialog] = useState(false);
 
   useEffect(() => {
     if (resource === ACTIVITY_RESOURCE.workOrder) {
@@ -243,7 +247,7 @@ const Diagram = ({
           message: data.message
         });
         fetchData();
-        setAttachemntDialog({ open: false, id: null, isClone: false });
+        setAttachemntDialog({ open: false, file: null, isClone: false });
         setFullScreen(false);
         setLoading(false);
       })
@@ -254,6 +258,22 @@ const Diagram = ({
   };
 
   const isSelectedAttachmentImage = checkImageType(selectedAttachment?.url?.split('.')[1]);
+
+  const handleRequestReject = (id) => {
+    axiosInstance()
+      .put('/attachment/deleteRequest', { _id: id, type: 'cancel' })
+      .then(({ data }) => {
+        fetchData();
+        toastConfig.setToastConfig({
+          open: true,
+          type: 'success',
+          message: data.message
+        });
+      })
+      .catch((error) => {
+        toastConfig.setToastConfig(error);
+      });
+  };
 
   return (
     <Box>
@@ -290,7 +310,7 @@ const Diagram = ({
               <ThemeButton
                 buttonType="theme"
                 onClick={() => {
-                  setAttachemntDialog({ open: true, id: null, isClone: false });
+                  setAttachemntDialog({ open: true, file: null, isClone: false });
                 }}
                 iconForMobile={<Add />}
                 mobileTooltip="Add"
@@ -324,57 +344,113 @@ const Diagram = ({
                           <span className="p-1">{expended[file?._id] ? <ExpandMoreIcon /> : <KeyboardArrowRight />}</span>
                           <Box ml={2}>
                             <Typography style={{ fontWeight: 600 }} className=" break-all" title={file?.name}>
-                              {file?.name} {`(${file?.createdBy?.user?.concatedName})`}
+                              {file?.name}
                             </Typography>
                           </Box>
                         </div>
-                        {!disableEdit && (
-                          <div className="flex gap-2">
-                            <HtmlTooltip title="Edit" placement="top" arrow>
-                              <IconButton
-                                size="small"
-                                color="inherit"
-                                aria-label="edit"
-                                disabled={file?.canEdit}
-                                onClick={(e) => {
-                                  e.stopPropagation();
-                                  setAttachemntDialog({ open: true, id: file?._id, isClone: false });
+                        <div className="flex gap-2">
+                          {!isEmpty(file?.deleteRequest) && file?.createdBy?.user?._id === user?._id && (
+                            <div className="flex gap-2">
+                              <ThemeButton
+                                buttonType="theme"
+                                onClick={() => {
+                                  handleDeleteFile([file._id]);
                                 }}
                               >
-                                <EditIcon style={{ fontSize: '18px' }} color={'primary'} />
-                              </IconButton>
-                            </HtmlTooltip>
-                            <HtmlTooltip title="Clone" placement="top" arrow>
-                              <IconButton
-                                size="small"
-                                color="inherit"
-                                aria-label="clone"
-                                onClick={(e) => {
-                                  e.stopPropagation();
-                                  setAttachemntDialog({ open: true, id: file?._id, isClone: true });
+                                Approve
+                              </ThemeButton>
+                              <ThemeButton
+                                buttonType="yellow"
+                                onClick={() => {
+                                  handleRequestReject(file._id);
                                 }}
                               >
-                                <FileCopyIcon style={{ fontSize: '18px' }} />
-                              </IconButton>
-                            </HtmlTooltip>
-                            <HtmlTooltip title="Delete" placement="top" arrow>
-                              <IconButton
-                                size="small"
-                                color="inherit"
-                                style={{ color: 'red' }}
-                                aria-label="delete"
-                                disabled={user?._id === file?.createdBy?.user?._id ? false : true}
-                                onClick={(e) => {
-                                  e.stopPropagation();
-                                  setSelectedFile(file);
-                                  setShowConfirmBox(true);
-                                }}
+                                Reject
+                              </ThemeButton>
+                            </div>
+                          )}
+                          <HtmlTooltip
+                            title={
+                              <div className="flex flex-col p-2">
+                                <p>
+                                  Uploaded By: <span>{file?.createdBy?.user?.concatedName}</span>
+                                </p>
+                                <p>
+                                  Uploaded Date: <span>{displayDate(file?.createdBy?.date)}</span>
+                                </p>
+                                {!isEmpty(file?.deleteRequest) && (
+                                  <>
+                                    <Divider></Divider>
+                                    <p>Delete Request Sent</p>
+                                    <p>
+                                      By: <span>{file?.deleteRequest?.user?.concatedName}</span>
+                                    </p>
+                                    <p>
+                                      Date: <span>{displayDate(file?.deleteRequest?.date)}</span>
+                                    </p>
+                                    <p>
+                                      Reason: <span>{file?.deleteRequest?.comment}</span>
+                                    </p>
+                                  </>
+                                )}
+                              </div>
+                            }
+                          >
+                            <IconButton size="small" color="inherit">
+                              <InfoOutlinedIcon style={{ fontSize: '18px' }} />
+                            </IconButton>
+                          </HtmlTooltip>
+                          {!disableEdit && (
+                            <>
+                              <HtmlTooltip title="Edit" placement="top" arrow>
+                                <IconButton
+                                  size="small"
+                                  color="inherit"
+                                  aria-label="edit"
+                                  disabled={!file?.canEdit}
+                                  onClick={(e) => {
+                                    e.stopPropagation();
+                                    setAttachemntDialog({ open: true, file: file, isClone: false });
+                                  }}
+                                >
+                                  <EditIcon style={{ fontSize: '18px' }} color={'primary'} />
+                                </IconButton>
+                              </HtmlTooltip>
+                              <HtmlTooltip title="Clone" placement="top" arrow>
+                                <IconButton
+                                  size="small"
+                                  color="inherit"
+                                  aria-label="clone"
+                                  onClick={(e) => {
+                                    e.stopPropagation();
+                                    setAttachemntDialog({ open: true, file: file, isClone: true });
+                                  }}
+                                >
+                                  <FileCopyIcon style={{ fontSize: '18px' }} />
+                                </IconButton>
+                              </HtmlTooltip>
+                              <HtmlTooltip
+                                title={!isEmpty(file?.deleteRequest) ? `Delete request sent to ${file?.createdBy?.user?.concatedName}` : 'Delete'}
+                                placement="top"
+                                arrow
                               >
-                                <Delete style={{ fontSize: '18px' }} color={user?._id === file?.createdBy?.user?._id ? 'error' : 'disabled'} />
-                              </IconButton>
-                            </HtmlTooltip>
-                          </div>
-                        )}
+                                <IconButton
+                                  size="small"
+                                  color="inherit"
+                                  aria-label="delete"
+                                  disabled={!isEmpty(file?.deleteRequest) ? true : false}
+                                  onClick={(e) => {
+                                    e.stopPropagation();
+                                    setSelectedFile(file);
+                                    setShowConfirmBox(true);
+                                  }}
+                                >
+                                  <Delete style={{ fontSize: '18px' }} color={!isEmpty(file?.deleteRequest) ? 'disabled' : 'error'} />
+                                </IconButton>
+                              </HtmlTooltip>
+                            </>
+                          )}
+                        </div>
                       </div>
                       <Collapse in={expended[file?._id]}>
                         <div className="border-t ">
@@ -482,17 +558,18 @@ const Diagram = ({
           maxWidth={'md'}
           onClose={(e, reason) => {
             if (reason !== 'backdropClick') {
-              setAttachemntDialog({ open: false, id: null, isClone: false });
+              setAttachemntDialog({ open: false, file: null, isClone: false });
             }
             setFullScreen(false);
           }}
           fullWidth
         >
           <ManageAttachment
-            attachmentId={attachemntDialog.id}
+            attachmentId={attachemntDialog.file?._id}
+            attachmentData={attachemntDialog.file}
             isClone={attachemntDialog.isClone}
             handleClose={() => {
-              setAttachemntDialog({ open: false, id: null, isClone: false });
+              setAttachemntDialog({ open: false, file: null, isClone: false });
               setFullScreen(false);
             }}
             relatedTo={getRelatedTo()}
@@ -503,7 +580,8 @@ const Diagram = ({
             showManimizeMaximize={true}
             fetchData={fetchData}
             attachmentType={attachmentType}
-            customhandleAdd={resource === ACTIVITY_RESOURCE.workOrder && !attachemntDialog.id && !showMaterialFilter ? customhandleAdd : null}
+            customhandleAdd={resource === ACTIVITY_RESOURCE.workOrder && !attachemntDialog.file && !showMaterialFilter ? customhandleAdd : null}
+            isOwner={attachemntDialog.file ? attachemntDialog.file?.createdBy?.user?._id === user?._id : true}
           />
         </Dialog>
       )}
@@ -515,8 +593,22 @@ const Diagram = ({
             setShowConfirmBox(false);
           }}
           onOk={() => {
-            handleDeleteFile([selectedFile._id]);
+            if (selectedFile?.createdBy?.user?._id === user?._id) {
+              handleDeleteFile([selectedFile._id]);
+            } else {
+              setCommentDialog(true);
+            }
             setShowConfirmBox(false);
+          }}
+        />
+      )}
+      {commentDialog && (
+        <Comment
+          onClose={() => setCommentDialog(false)}
+          file={selectedFile}
+          onSuccess={() => {
+            setCommentDialog(false);
+            fetchData();
           }}
         />
       )}
